backend/requirements.txt
# PM Architect Backend - Lite Orchestrator
# Minimal dependencies for production deployment

# FastAPI framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Pydantic for data validation
pydantic==2.9.2
pydantic-settings==2.6.1

# Google Gemini AI SDK
google-generativeai==0.8.3

# HTTP client for API calls
httpx==0.27.2

# Environment variable management
python-dotenv==1.0.1

# File upload support (optional, for future features)
python-multipart==0.0.20
